tools: Cleaned up to pass /W4 without warnings.

+++ tests/gpu/tiling/vulkan_helpers.cpp
@@ -8,7 +8,7 @@ vk::Instance gVulkan = {};
 vk::PhysicalDevice gPhysDevice = {};
 vk::Device gDevice = {};
 vk::Queue gQueue = {};
-uint32_t gQueueFamilyIndex = -1;
+uint32_t gQueueFamilyIndex = static_cast<uint32_t>(-1);
 vk::CommandPool gCommandPool = {};
 
 static VKAPI_ATTR VkBool32 VKAPI_CALL
